# sistemareservas/src/swagger.yaml
openapi: 3.0.1
info:
  title: Sistema de Reservas API
  description: API REST para gestionar usuarios, reservas, recursos y pagos.
  version: 1.0.0
servers:
  - url: http://localhost:8080

paths:

  /usuarios:
    get:
      summary: Listar usuarios
      responses:
        '200':
          description: Lista de usuarios
    post:
      summary: Crear usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Usuario'
      responses:
        '200':
          description: Usuario creado

  /usuarios/{id}:
    get:
      summary: Obtener usuario por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Usuario encontrado
    put:
      summary: Actualizar usuario
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Usuario'
      responses:
        '200':
          description: Usuario actualizado
    delete:
      summary: Eliminar usuario
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Usuario eliminado

  /reservas:
    get:
      summary: Listar reservas
      responses:
        '200':
          description: Lista de reservas
    post:
      summary: Crear reserva
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reserva'
      responses:
        '200':
          description: Reserva creada

  /reservas/{id}:
    get:
      summary: Obtener reserva por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Reserva encontrada
    put:
      summary: Actualizar reserva
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reserva'
      responses:
        '200':
          description: Reserva actualizada
    delete:
      summary: Eliminar reserva
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Reserva eliminada

  /reservas/{id}/total:
    get:
      summary: Obtener total de reserva
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Total calculado

  /recursos:
    get:
      summary: Listar recursos
      responses:
        '200':
          description: Lista de recursos
    post:
      summary: Crear recurso
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Recurso'
      responses:
        '200':
          description: Recurso creado

  /recursos/{id}:
    get:
      summary: Obtener recurso por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Recurso encontrado
    put:
      summary: Actualizar recurso
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Recurso'
      responses:
        '200':
          description: Recurso actualizado
    delete:
      summary: Eliminar recurso
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Recurso eliminado

  /pagos:
    get:
      summary: Listar pagos
      responses:
        '200':
          description: Lista de pagos
    post:
      summary: Crear pago
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pago'
      responses:
        '200':
          description: Pago creado

  /pagos/{id}:
    get:
      summary: Obtener pago por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Pago encontrado
    put:
      summary: Actualizar pago
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pago'
      responses:
        '200':
          description: Pago actualizado
    delete:
      summary: Eliminar pago
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Pago eliminado

components:
  schemas:
    Usuario:
      type: object
      properties:
        idUsuario:
          type: integer
        nombreCompleto:
          type: string
        numTelefono:
          type: integer
        email:
          type: string
        role:
          type: string
        password:
          type: string

    Reserva:
      type: object
      properties:
        idReserva:
          type: integer
        usuario:
          $ref: '#/components/schemas/Usuario'
        recursos:
          type: array
          items:
            $ref: '#/components/schemas/Recurso'
        fechaInicio:
          type: string
          format: date-time
        fechaFin:
          type: string
          format: date-time
        estadoReservacion:
          type: string

    Recurso:
      type: object
      properties:
        idRecurso:
          type: integer
        nombre:
          type: string
        precio:
          type: number
          format: double

    Pago:
      type: object
      properties:
        idPago:
          type: integer
        monto:
          type: number
          format: double
        metodoPago:
          type: string
        fechaPago:
          type: string
          format: date-time
        reserva:
          $ref: '#/components/schemas/Reserva'